+++ clients/web/src/lib/browserTabVisibility.test.ts
