+++ clients/web/src/lib/browserTabVisibility.ts
