+++ clients/web/src/utils/sandbox-csp.test.ts
