+++ clients/web/src/utils/sandbox-csp.ts
